(maint) Remove dujour whitelist testing
As of PE-24341 it does not send a meaningful whitelist (it
sends nil so that it will be ignored) and in PE-24341 database access
was removed in its entirety. Consequently testing whitelist
functionality can't be tested at the dujour integration level.

This patch removes those dujour integration tests and depends on the
whitelist testing in core-test namespace. It also removes all of the
db/postgres related testing setup that was in service of those tests.